refactor(grading): relocate pure grading read-model code out of services
Phase 2 of fixing the repository -> services inversions (phase 1 moved
aggregate_grading_status to utils/grading_status.py).

course_member_grading_read.py was a pure read-model *builder* -- it takes a
CourseMemberGradingReadBackend Protocol and imports only DTOs + the pure
process_hierarchical_stats; no DB/IO. It was merely filed under services/.
course_member_grading_stats.py is likewise pure aggregation. Move both down so
the view repository no longer imports upward:

- services/course_member_grading_stats.py -> utils/grading_stats.py   (pure)
- services/course_member_grading_read.py  -> repositories/grading_read.py (read model)

Importers updated: course_member_gradings_view (now a peer import -- inversion
gone), analytics/{service,grading_repository}, business_logic/course_member_gradings,
and the analytics-grading-backend test (23 passed).

Also documents the remaining view_mappers -> services.result_storage dependency
as deliberate: the view_*/mapper modules are a read-model (query) layer that may
orchestrate service reads (MinIO result JSON/artifacts), not the persistence
repository layer -- so it is not an inverted dependency.
